Complete implementation of D3D12 PSO XML parser

+++ lexgine/core/dx/d3d12/d3d12_pso_xml_parser.h
@@ -65,10 +65,10 @@ public:
 
 
     //! returns total amount of parsed graphics PSO descriptors
-    uint32_t getNumberOfParsedGraphicsPSOs() const;
+    size_t getNumberOfParsedGraphicsPSOs() const;
 
     //! returns total amount of parsed computed PSO descriptors
-    uint32_t getNumberOfParsedComputePSOs() const;
+    size_t getNumberOfParsedComputePSOs() const;
 
 private:
     class impl;
